+++ test/src/response/filtered_stream_response_test.dart
@@ -38,7 +38,8 @@ void main() {
         'attachments: null, privateMetrics: null, organicMetrics: null, '
         'promotedMetrics: null, publicMetrics: null, geo: null, lang: null, '
         'isPossiblySensitive: null, replySetting: null, source: null, '
-        'withheld: null, createdAt: null), includes: null, meta: null, '
+        'editControls: null, editHistoryTweetIds: null, withheld: null, '
+        'createdAt: null), includes: null, meta: null, '
         'matchingRules: [MatchingRule(id: 456, tag: test)])');
   });
 }