+++ test/src/service/tweets/tweets_service_test.dart
@@ -12,6 +12,7 @@ import 'package:twitter_api_core/src/exception/twitter_exception.dart';
 import 'package:twitter_api_v2/src/response/filtered_stream_response.dart';
 import 'package:twitter_api_v2/src/response/twitter_response.dart';
 import 'package:twitter_api_v2/src/response/twitter_stream_response.dart';
+import 'package:twitter_api_v2/src/service/tweets/decahose_partition.dart';
 import 'package:twitter_api_v2/src/service/tweets/exclude_tweet_type.dart';
 import 'package:twitter_api_v2/src/service/tweets/filtering_rule_data.dart';
 import 'package:twitter_api_v2/src/service/tweets/filtering_rule_meta.dart';
@@ -1014,6 +1015,14 @@ void main() {
       expect(response, isA<TwitterResponse>());
       expect(response.data, isA<TweetData>());
       expect(response.data.id, '1067094924124872705');
+
+      final editControls = response.data.editControls!;
+
+      expect(editControls.toJson(), {
+        'is_edit_eligible': true,
+        'edits_remaining': '4',
+        'editable_until': '2022-08-21T09:35:20.311'
+      });
     });
 
     test('with expansions', () async {
@@ -2307,17 +2316,17 @@ void main() {
     });
   });
 
-  group('.connectVolumeStream', () {
+  group('.connectSampleStream', () {
     test('normal case', () async {
       final tweetsService = TweetsService(
         context: context.buildSendStub(
           UserContext.oauth2Only,
-          'test/src/service/tweets/data/connect_volume_stream.json',
+          'test/src/service/tweets/data/connect_sample_stream.json',
           const {'backfill_minutes': '5'},
         ),
       );
 
-      final response = await tweetsService.connectVolumeStream(
+      final response = await tweetsService.connectSampleStream(
         backfillMinutes: 5,
       );
 
@@ -2333,12 +2342,70 @@ void main() {
       final tweetsService = TweetsService(
         context: context.buildSendStub(
           UserContext.oauth2Only,
-          'test/src/service/tweets/data/connect_volume_stream_with_error.json',
+          'test/src/service/tweets/data/connect_sample_stream_with_error.json',
           const {'backfill_minutes': '5'},
         ),
       );
 
-      final response = await tweetsService.connectVolumeStream(
+      final response = await tweetsService.connectSampleStream(
+        backfillMinutes: 5,
+      );
+
+      final data = <TwitterResponse<TweetData, void>>[];
+      final errors = <dynamic>[];
+
+      final stream = response.stream;
+
+      await for (final event in stream.handleError(errors.add)) {
+        data.add(event);
+      }
+
+      expect(data.length, 5);
+      expect(errors.length, 1);
+      expect(errors.single, isA<TwitterException>());
+    });
+  });
+
+  group('.connectSample10Stream', () {
+    test('normal case', () async {
+      final tweetsService = TweetsService(
+        context: context.buildSendStub(
+          UserContext.oauth2Only,
+          'test/src/service/tweets/data/connect_sample10_stream.json',
+          const {
+            'partition': '1',
+            'backfill_minutes': '5',
+          },
+        ),
+      );
+
+      final response = await tweetsService.connectSample10Stream(
+        partition: DecahosePartition.section1,
+        backfillMinutes: 5,
+      );
+
+      final data = await response.stream.toList();
+
+      expect(response,
+          isA<TwitterStreamResponse<TwitterResponse<TweetData, void>>>());
+      expect(data, isA<List<TwitterResponse<TweetData, void>>>());
+      expect(data.length, 70);
+    });
+
+    test('with error', () async {
+      final tweetsService = TweetsService(
+        context: context.buildSendStub(
+          UserContext.oauth2Only,
+          'test/src/service/tweets/data/connect_sample10_stream_with_error.json',
+          const {
+            'partition': '2',
+            'backfill_minutes': '5',
+          },
+        ),
+      );
+
+      final response = await tweetsService.connectSample10Stream(
+        partition: DecahosePartition.section2,
         backfillMinutes: 5,
       );