+++ CHANGELOG.md
@@ -1,8 +1,25 @@
 # Release Note
 
+## v4.2.0
+
+- Deprecated properties in `TwitterApi` below and added new properties. It will be removed in v5.0.0, so please replace them to new property names. ([#483](https://github.com/twitter-dart/twitter-api-v2/issues/483))
+  - From `tweetsService` to `tweets`
+  - From `usersService` to `users`
+  - From `spacesService` to `spaces`
+  - From `listsService` to `lists`
+  - From `mediaService` to `media`
+  - From `complianceService` to `compliance`
+
+## v4.1.5
+
+- Added `TweetEditControls` object. This object holds information about the editing of tweets. You can use it from `TweetData.editControls`, and also you must pass the `editControls` element to `tweetFields`. ([#472](https://github.com/twitter-dart/twitter-api-v2/issues/472))
+- Added `editHistoryTweetIds` field in `TweetData`. ([#471](https://github.com/twitter-dart/twitter-api-v2/issues/471))
+
 ## v4.1.4
 
 - Renamed object from `ComplianceData` to `BatchComplianceData`. If you are using the object name `ComplianceData` explicitly, modify it to `BatchComplianceData`. ([#462](https://github.com/twitter-dart/twitter-api-v2/issues/462))
+- Deprecated `connectVolumeStream` and added `connectSampleStream` instead. Please replace method name from `connectVolumeStream` to `connectSampleStream`. ([#479](https://github.com/twitter-dart/twitter-api-v2/issues/479))
+- Supported `connectSample10Stream` in Tweets Service. This feature is available to Enterprise levels of access. Also, this is commonly referred to as the **Decahose**. ([#473](https://github.com/twitter-dart/twitter-api-v2/issues/473))
 
 ## v4.1.3